requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
numpy==1.26.0
